fix(form-builder): only render `<span>` inside `<div>`
